% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{fallback_global_quantile}
\alias{fallback_global_quantile}
\title{Calculate Global Quantiles for Fallback}
\usage{
fallback_global_quantile(y_train, alpha)
}
\arguments{
\item{y_train}{A numeric vector of the response variable from the training data.}

\item{alpha}{A numeric value between 0 and 1, representing the significance
level for the prediction interval.}
}
\value{
A list containing two components:
  \item{L}{The lower global quantile.}
  \item{U}{The upper global quantile.}
}
\description{
This internal utility function calculates global empirical quantiles from
the training response variable `y_train`. These quantiles serve as a
fallback mechanism when local estimates are deemed unreliable.
}
