% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinp_api.R
\name{fit_pinp}
\alias{fit_pinp}
\alias{fit_pinp.default}
\alias{fit_pinp.formula}
\alias{fit_pinp.data.frame}
\title{Fit a Nonparametric Prediction Interval Model}
\usage{
fit_pinp(object, ...)

\method{fit_pinp}{default}(object, ...)

\method{fit_pinp}{formula}(formula, data, method = "knn_eqt", fallback_formula = NULL, ...)

\method{fit_pinp}{data.frame}(object, y, method = "knn_eqt", fallback_formula = NULL, ...)
}
\arguments{
\item{object}{An object to be used for fitting the model.}

\item{...}{Additional arguments passed to the underlying fitting function,
such as `k` for the kNN method.}

\item{formula}{A formula specifying the model, e.g., `y ~ x1 + x2`.}

\item{data}{A data frame containing the variables in the formula.}

\item{method}{A character string specifying the method to use. Default is "knn_eqt".}

\item{fallback_formula}{An optional formula to use for the fallback linear model.
If `NULL`, a formula `y ~ .` is constructed from the provided `X` and `y`.}

\item{y}{A numeric vector of the response variable.}
}
\value{
A fitted object of class `pinp_fit`.
}
\description{
`fit_pinp` is a generic function for fitting a model to be used for
generating nonparametric prediction intervals. The function invokes the
appropriate method based on the class of the input object.
}
