% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_numeric.R
\name{grid_y}
\alias{grid_y}
\title{Generate a Grid for the Response Variable}
\usage{
grid_y(y, len, pad)
}
\arguments{
\item{y}{A numeric vector representing the response variable.}

\item{len}{An integer specifying the desired length of the grid.}

\item{pad}{A numeric value specifying the padding to add to the range of `y`.
The grid will extend `pad` times the range of `y` on both sides.}
}
\value{
A numeric vector representing the generated grid for `y`.
}
\description{
This internal utility function generates a sequence of points (a grid)
for the response variable `y`. It is used to create evaluation points
for conditional density estimation.
}
