% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_knn_eqt.R
\name{knn_eqt_fit}
\alias{knn_eqt_fit}
\title{Fit kNN Equal-Tailed Prediction Interval Model}
\usage{
knn_eqt_fit(X, y, k = 50, ...)
}
\arguments{
\item{X}{A matrix or data frame of predictor variables from the training data.}

\item{y}{A numeric vector of the response variable from the training data.}

\item{k}{The number of neighbors. Defaults to 50.}

\item{...}{Additional arguments (not used in this method).}
}
\value{
A list representing the fitted kNN model, containing:
   \item{X_train}{The training predictor variables.}
   \item{y_train}{The training response variable.}
   \item{k}{The number of neighbors used.}
}
\description{
This internal function "fits" a kNN-based model for generating equal-tailed
prediction intervals. For kNN, fitting primarily involves storing the
training data and the number of neighbors `k`.
}
