% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_compare}
\alias{plot_compare}
\title{Compare Prediction Intervals of Two Models}
\usage{
plot_compare(
  model1,
  model2,
  data,
  n_grid = 100,
  ...,
  model1_name = "Model 1 (blue)",
  model2_name = "Model 2 (red)"
)
}
\arguments{
\item{model1}{The first fitted object of class `pinp_fit`.}

\item{model2}{The second fitted object of class `pinp_fit`.}

\item{data}{The training data frame used for plotting the points. The formula
from `model1` will be used to extract the correct columns.}

\item{n_grid}{An integer specifying the number of points for the prediction grid.}

\item{...}{Additional arguments passed to both `predict_intervals` calls.}

\item{model1_name}{A character string for the name of the first model in the plot title.}

\item{model2_name}{A character string for the name of the second model in the plot title.}
}
\value{
A `ggplot` object, which is also printed to the console.
}
\description{
Plots the prediction intervals from two different `pinp_fit` models on the
same data for visual comparison.
}
