% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinp_api.R
\name{predict_intervals}
\alias{predict_intervals}
\alias{predict_intervals.pinp_fit}
\title{Predict Nonparametric Intervals}
\usage{
predict_intervals(object, ...)

\method{predict_intervals}{pinp_fit}(object, new_data, alpha = 0.1, use_fallback = TRUE, ...)
}
\arguments{
\item{object}{A fitted object of class `pinp_fit`.}

\item{...}{Additional arguments to be passed to the specific method.}

\item{new_data}{A data frame containing the predictor variables for which
to generate prediction intervals.}

\item{alpha}{A numeric value between 0 and 1 representing the significance
level. The prediction interval will be of size `1 - alpha`.}

\item{use_fallback}{A logical value indicating whether to enable the
fallback mechanism. Defaults to `TRUE`.}
}
\value{
A data frame containing the predicted intervals and diagnostics.
}
\description{
`predict_intervals` is a generic function for predicting nonparametric
prediction intervals from a fitted `pinp` object.
}
