% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_neighbors.R
\name{query_knn}
\alias{query_knn}
\title{Query K-Nearest Neighbors}
\usage{
query_knn(X_train, X_new, k)
}
\arguments{
\item{X_train}{A matrix or data frame of training data features.}

\item{X_new}{A matrix or data frame of new data points for which to find
neighbors.}

\item{k}{The number of nearest neighbors to find.}
}
\value{
A list containing two components:
  \item{nn.index}{A matrix of indices of the nearest neighbors.}
  \item{nn.dist}{A matrix of distances to the nearest neighbors.}
}
\description{
This internal utility function finds the k-nearest neighbors in `X_train`
for each point in `X_new`. It serves as a wrapper around the `FNN::get.knnx`
function.
}
