% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sim_hetero}
\alias{sim_hetero}
\title{Simulate: Heteroskedastic Data}
\usage{
sim_hetero(n = 500, seed = 101)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Random seed}
}
\value{
A data.frame
}
\description{
Simulate: Heteroskedastic Data
}
