% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sim_mix}
\alias{sim_mix}
\title{Simulate: Multimodal (Mix) Data}
\usage{
sim_mix(n = 500, seed = 102)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Random seed}
}
\value{
A data.frame
}
\description{
Simulate: Multimodal (Mix) Data
}
