% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sim_skew}
\alias{sim_skew}
\title{Simulate: Skewed Data}
\usage{
sim_skew(n = 500, seed = 103)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Random seed}
}
\value{
A data.frame
}
\description{
Simulate: Skewed Data
}
