% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{sim_sparse}
\alias{sim_sparse}
\title{Simulate: Sparse Region Data (for robustness testing)}
\usage{
sim_sparse(n_dense = 500, n_sparse = 5, seed = 104)
}
\arguments{
\item{n_dense}{Number of points in the dense region}

\item{n_sparse}{Number of points in the sparse region}

\item{seed}{Random seed}
}
\value{
A data.frame
}
\description{
Data is very sparse in the X > 0.5 region.
}
