% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_numeric.R
\name{weighted_quantile}
\alias{weighted_quantile}
\title{Calculate Weighted Quantiles}
\usage{
weighted_quantile(values, weights, probs)
}
\arguments{
\item{values}{A numeric vector of values for which to calculate the quantiles.}

\item{weights}{A numeric vector of weights, of the same length as `values`.
All weights must be non-negative.}

\item{probs}{A numeric vector of probabilities with values in [0, 1].}
}
\value{
A numeric vector of quantiles corresponding to the given
  probabilities.
}
\description{
Computes quantiles for a set of values, taking into account a vector of
weights. This function is a robust implementation for calculating weighted
quantiles, which is particularly useful in nonparametric statistics.
}
\examples{
values <- c(1, 2, 3, 4, 5)
weights <- c(1, 1, 1, 1, 1)
weighted_quantile(values, weights, probs = c(0.25, 0.5, 0.75))

values <- c(10, 20, 30, 40, 50)
weights <- c(1, 2, 3, 2, 1)
weighted_quantile(values, weights, probs = 0.5)
}
