% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_kde_eqt.R
\name{kde_eqt_fit}
\alias{kde_eqt_fit}
\title{Fit KDE Equal-Tailed Prediction Interval Model}
\usage{
kde_eqt_fit(X, y, bw_x = NULL, bw_y = NULL, ...)
}
\arguments{
\item{X}{A matrix or data frame of predictor variables from the training data.}

\item{y}{A numeric vector of the response variable from the training data.}

\item{bw_x}{A numeric vector specifying the bandwidth for each predictor
variable. If `NULL`, bandwidths are calculated for each column of `X`
using `stats::bw.nrd0`.}

\item{bw_y}{A numeric scalar specifying the bandwidth for the response
variable. If `NULL`, it is calculated using `stats::bw.nrd0`.}

\item{...}{Additional arguments (not used in this method).}
}
\value{
A list representing the fitted KDE model.
}
\description{
This internal function "fits" a KDE-based model for generating equal-tailed
prediction intervals. It handles multi-dimensional predictor variables by
preparing to use a product kernel.
}
